Move rate limiter to middleware
